+++ neuralcompression/layers/__init__.py
@@ -5,5 +5,5 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 """
 
-from ._non_negative_parameter import NonNegativeParameter
+from ._non_negative_parameterization import NonNegativeParameterization
 from .gdn import SimplifiedGDN, SimplifiedInverseGDN
