+++ tests/layers/test_non_negative_parameterization.py
@@ -1,7 +1,14 @@
+"""
+Copyright (c) Facebook, Inc. and its affiliates.
+
+This source code is licensed under the MIT license found in the
+LICENSE file in the root directory of this source tree.
+"""
+
 import torch
 import torch.testing
 
-from neuralcompression.layers import NonNegativeParameter
+from neuralcompression.layers import NonNegativeParameterization
 
 
 class TestNonNegativeParameter:
